clicking & eventListener on save btn working
Clicking now disables wordFinder when tippy is showing, and reactivates it when it is hidden. This makes it such that you can click anywhere on screen to minimize tippy and it will not try to evaluate tippy contents when tippy is showing. Clicking on "Save This Word" in tippy gives us access to t and translation.